requirements.txt
aiofiles==25.1.0
aiohappyeyeballs==2.6.1
aiohttp==3.13.1
aiosignal==1.4.0
asgiref==3.10.0
asn1crypto==1.5.1
attrs==25.4.0
cachetools==6.2.1
certifi==2025.10.5
cffi==2.0.0
charset-normalizer==3.4.4
cloud-sql-python-connector==1.18.5
cryptography==46.0.3
Django==5.2.7
dnspython==2.8.0
frozenlist==1.8.0
google-api-core==2.27.0
google-auth==2.41.1
google-cloud-secret-manager==2.25.0
googleapis-common-protos==1.71.0
grpc-google-iam-v1==0.14.3
grpcio==1.76.0
grpcio-status==1.76.0
gunicorn==23.0.0
idna==3.11
multidict==6.7.0
numpy
packaging==25.0
pandas
pg8000==1.31.5
propcache==0.4.1
proto-plus==1.26.1
protobuf==6.33.0
psycopg2-binary==2.9.11
pyasn1==0.6.1
pyasn1_modules==0.4.2
pycparser==2.23
python-dateutil==2.9.0.post0
python-dotenv==1.2.1
requests==2.32.5
rsa==4.9.1
scipy
scramp==1.4.6
six==1.17.0
sqlparse==0.5.3
typing_extensions==4.15.0
urllib3==2.5.0
yarl==1.22.0
